chore: clean up qr dqn
